split homeassistant to new class
